refactor: reorganiza estrutura de pacotes do projeto
